refactor: move db initialization to the struct
